requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-jose[cryptography]>=3.3.0
pgvector
email-validator
openai>=1.0.0
python-dotenv>=1.0.0
greenlet
